refactoring and fixing datasource selector button
